requirements.txt
fastapi
psycopg2-binary
python-dotenv
uvicorn[standard]